- Add edit button for profile page - Change Buttons for registration and edit profile pages - refactor user's profile

+++ src/Connection/AdminBundle/Entity/Background.php
@@ -130,6 +130,8 @@ class Background
      * @ORM\PreRemove()
      */
     public function removeFiles() {
-        unlink($this->getUploadRootDir());
+        if ( file_exists($this->getUploadRootDir()) ) {
+            unlink($this->getUploadRootDir());
+        }
     }
 }
